+++ app/serializers/v2/event.ts
@@ -1,8 +1,9 @@
 import { dbAdapter } from '../../models';
 import { Nullable, UUID } from '../../support/types';
 import type { EventRecord } from '../../support/DbAdapter';
+import { HIDDEN_CREATOR_EVENT_TYPES } from '../../support/EventTypes';
 
-import { userSerializerFunction } from './user';
+import { serializeUsersByIds } from './user';
 
 export async function serializeEvents(events: EventRecord[], viewerId: Nullable<UUID> = null) {
   const accountIntIds = new Set<Nullable<number>>();
@@ -10,6 +11,13 @@ export async function serializeEvents(events: EventRecord[], viewerId: Nullable<
   const commentIntIds = new Set<Nullable<number>>();
 
   for (const event of events) {
+    if (
+      HIDDEN_CREATOR_EVENT_TYPES.includes(event.event_type) &&
+      event.user_id === event.target_user_id
+    ) {
+      event.created_by_user_id = null;
+    }
+
     accountIntIds.add(event.user_id);
     accountIntIds.add(event.target_user_id);
     accountIntIds.add(event.created_by_user_id);
@@ -74,22 +82,11 @@ export async function serializeEvents(events: EventRecord[], viewerId: Nullable<
 
   // Now collecting user information for the output
   const accountIds = accountIdRows.map((r) => r.uid);
-  const allGroupAdmins = await dbAdapter.getGroupsAdministratorsIds(accountIds, viewerId);
-  Object.values(allGroupAdmins).forEach((ids) =>
-    ids.forEach((s) => !accountIds.includes(s) && accountIds.push(s)),
-  );
-
-  const [allUsersAssoc, allStatsAssoc] = await Promise.all([
-    dbAdapter.getUsersByIdsAssoc(accountIds),
-    dbAdapter.getUsersStatsAssoc(accountIds),
-  ]);
-
-  const serializeUser = userSerializerFunction(allUsersAssoc, allStatsAssoc, allGroupAdmins);
-  const accounts = Object.keys(allUsersAssoc).map(serializeUser);
+  const sAccounts = await serializeUsersByIds(accountIds, viewerId);
 
   return {
     events: serializedEvents,
-    users: accounts.filter((u) => u.type === 'user'),
-    groups: accounts.filter((u) => u.type === 'group'),
+    users: sAccounts.filter((u) => u.type === 'user'),
+    groups: sAccounts.filter((u) => u.type === 'group'),
   };
 }